requirements.txt
requests
ipwhois
python-dotenv
Flask
slack_sdk
slackeventsapi